+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Support reset NEC display settings as means of recovery from breaking external changes
 - Internal client-server communication re-implemented using websocket instead of HTTP requests
 - Manage compute module fan based on cpu temperature in NEC displays
+- Support Network set manual & DHCP
 
 ### Fixed
 - Casually not loading of applet after restart app